[feature](shared/api): DMS API classification
