+++ packages/shared/lib/api/base/service/BasicInfo/index.d.ts
@@ -1,0 +1,11 @@
+import { IGetBasicInfoReply, IGenericResp } from '../common.d';
+
+export interface IGetBasicInfoReturn extends IGetBasicInfoReply {}
+
+export interface IPersonalizationParams {
+  title?: string;
+
+  file?: any;
+}
+
+export interface IPersonalizationReturn extends IGenericResp {}
