+++ packages/shared/lib/api/base/service/CompanyNotice/index.d.ts
@@ -1,0 +1,11 @@
+import {
+  IGetCompanyNoticeReply,
+  IUpdateCompanyNoticeReq,
+  IGenericResp
+} from '../common.d';
+
+export interface IGetCompanyNoticeReturn extends IGetCompanyNoticeReply {}
+
+export interface IUpdateCompanyNoticeParams extends IUpdateCompanyNoticeReq {}
+
+export interface IUpdateCompanyNoticeReturn extends IGenericResp {}
