+++ packages/shared/lib/api/base/service/DBService/index.enum.ts
@@ -1,0 +1,13 @@
+/* tslint:disable no-duplicate-string */
+
+export enum ListDBServicesOrderByEnum {
+  'name' = 'name'
+}
+
+export enum ListDBServiceTipsFunctionalModuleEnum {
+  'save_audit_plan' = 'save_audit_plan',
+
+  'create_workflow' = 'create_workflow',
+
+  'create_export_task' = 'create_export_task'
+}
