+++ packages/shared/lib/api/base/service/DMSPlugin/index.d.ts
@@ -1,0 +1,5 @@
+import { IRegisterDMSPluginReq, IGenericResp } from '../common.d';
+
+export interface IRegisterDMSPluginParams extends IRegisterDMSPluginReq {}
+
+export interface IRegisterDMSPluginReturn extends IGenericResp {}
