+++ packages/shared/lib/api/base/service/DMSProxy/index.d.ts
@@ -1,0 +1,6 @@
+import { IRegisterDMSProxyTargetReq, IGenericResp } from '../common.d';
+
+export interface IRegisterDMSProxyTargetParams
+  extends IRegisterDMSProxyTargetReq {}
+
+export interface IRegisterDMSProxyTargetReturn extends IGenericResp {}
