+++ packages/shared/lib/api/base/service/DataExportWorkflows/index.enum.ts
@@ -1,0 +1,17 @@
+/* tslint:disable no-duplicate-string */
+
+export enum ListDataExportWorkflowsFilterByStatusEnum {
+  'wait_for_approve' = 'wait_for_approve',
+
+  'wait_for_export' = 'wait_for_export',
+
+  'exporting' = 'exporting',
+
+  'rejected' = 'rejected',
+
+  'cancel' = 'cancel',
+
+  'failed' = 'failed',
+
+  'finish' = 'finish'
+}
