+++ packages/shared/lib/api/base/service/Masking/index.d.ts
@@ -1,0 +1,3 @@
+import { IListMaskingRulesReply } from '../common.d';
+
+export interface IListMaskingRulesReturn extends IListMaskingRulesReply {}
