+++ packages/shared/lib/api/base/service/Member/index.enum.ts
@@ -1,0 +1,5 @@
+/* tslint:disable no-duplicate-string */
+
+export enum ListMembersOrderByEnum {
+  'user_uid' = 'user_uid'
+}
