+++ packages/shared/lib/api/base/service/MemberGroup/index.enum.ts
@@ -1,0 +1,5 @@
+/* tslint:disable no-duplicate-string */
+
+export enum ListMemberGroupsOrderByEnum {
+  'name' = 'name'
+}
