+++ packages/shared/lib/api/base/service/Notification/index.d.ts
@@ -1,0 +1,5 @@
+import { INotification, INotificationReply } from '../common.d';
+
+export interface INotificationParams extends INotification {}
+
+export interface INotificationReturn extends INotificationReply {}
