+++ packages/shared/lib/api/base/service/OpPermission/index.enum.ts
@@ -1,0 +1,13 @@
+/* tslint:disable no-duplicate-string */
+
+export enum ListOpPermissionsOrderByEnum {
+  'name' = 'name'
+}
+
+export enum ListOpPermissionsFilterByTargetEnum {
+  'all' = 'all',
+
+  'user' = 'user',
+
+  'member' = 'member'
+}
