+++ packages/shared/lib/api/base/service/Project/index.enum.ts
@@ -1,0 +1,9 @@
+/* tslint:disable no-duplicate-string */
+
+export enum ListProjectsOrderByEnum {
+  'name' = 'name'
+}
+
+export enum ExportProjectsOrderByEnum {
+  'name' = 'name'
+}
