+++ packages/shared/lib/api/base/service/Role/index.enum.ts
@@ -1,0 +1,5 @@
+/* tslint:disable no-duplicate-string */
+
+export enum ListRolesOrderByEnum {
+  'name' = 'name'
+}
