+++ packages/shared/lib/api/base/service/Session/index.d.ts
@@ -1,0 +1,18 @@
+import {
+  IAddSessionReq,
+  IAddSessionReply,
+  IGenericResp,
+  IGetUserBySessionReply
+} from '../common.d';
+
+export interface IAddSessionParams extends IAddSessionReq {}
+
+export interface IAddSessionReturn extends IAddSessionReply {}
+
+export interface IDelSessionReturn extends IGenericResp {}
+
+export interface IGetUserBySessionParams {
+  user_uid?: string;
+}
+
+export interface IGetUserBySessionReturn extends IGetUserBySessionReply {}
