+++ packages/shared/lib/api/base/service/Webhook/index.d.ts
@@ -1,0 +1,5 @@
+import { IWebHookSendMessageReq, IWebHookSendMessageReply } from '../common.d';
+
+export interface IWebHookSendMessageParams extends IWebHookSendMessageReq {}
+
+export interface IWebHookSendMessageReturn extends IWebHookSendMessageReply {}
