--- packages/shared/lib/api/base/service/dms/index.enum.ts
@@ -1,69 +1,0 @@
-/* tslint:disable no-duplicate-string */
-
-export enum ListOpPermissionsOrderByEnum {
-  'name' = 'name'
-}
-
-export enum ListOpPermissionsFilterByTargetEnum {
-  'all' = 'all',
-
-  'user' = 'user',
-
-  'member' = 'member'
-}
-
-export enum ListProjectsOrderByEnum {
-  'name' = 'name'
-}
-
-export enum ExportProjectsOrderByEnum {
-  'name' = 'name'
-}
-
-export enum ListDataExportWorkflowsFilterByStatusEnum {
-  'wait_for_approve' = 'wait_for_approve',
-
-  'wait_for_export' = 'wait_for_export',
-
-  'exporting' = 'exporting',
-
-  'rejected' = 'rejected',
-
-  'cancel' = 'cancel',
-
-  'failed' = 'failed',
-
-  'finish' = 'finish'
-}
-
-export enum ListDBServicesOrderByEnum {
-  'name' = 'name'
-}
-
-export enum ListDBServiceTipsFunctionalModuleEnum {
-  'save_audit_plan' = 'save_audit_plan',
-
-  'create_workflow' = 'create_workflow',
-
-  'create_export_task' = 'create_export_task'
-}
-
-export enum ListMemberGroupsOrderByEnum {
-  'name' = 'name'
-}
-
-export enum ListMembersOrderByEnum {
-  'user_uid' = 'user_uid'
-}
-
-export enum ListRolesOrderByEnum {
-  'name' = 'name'
-}
-
-export enum ListUserGroupsOrderByEnum {
-  'name' = 'name'
-}
-
-export enum ListUsersOrderByEnum {
-  'name' = 'name'
-}
